+++ src/almost_linear/almost_linear_time.hpp
@@ -10,20 +10,20 @@
 
 class AlmostLinearTime {
 public:
-    explicit AlmostLinearTime(const std::map<std::pair<int,int>, int>& graph);
+    explicit AlmostLinearTime(const std::map<std::pair<int,int>, long long>& graph);
 
     // Ricerca binaria sul valore ottimale — restituisce il valore del max-flow
-    int max_flow(int source, int sink);
+    long long max_flow(int source, int sink);
 
     // Risolve il max-flow con una stima fissata del valore ottimale
-    std::pair<int, Eigen::VectorXd> max_flow_with_guess(
+    std::pair<long long, Eigen::VectorXd> max_flow_with_guess(
         int source,
         int sink,
-        int optimal_flow,
-        const std::vector<int>* lower_capacities = nullptr
+        long long optimal_flow,
+        const std::vector<long long>* lower_capacities = nullptr
     );
 
 private:
     std::vector<std::pair<int,int>> _edges;
-    std::vector<int>                _capacities;
+    std::vector<long long>  _capacities;
 };
